Fix a bug when using 1.9.2
Trying to #flatten an array of tasks caused errors in 1.9.2

Basically, Array#flatten in 1.9.2 calls #to_ary on each child element
and if a NoMethodError is raised, it assumes that the element is not
an array that needs to be flattened.

In 1.8.7, Array#flatten first called #respond_to? to determine if the
object had a #to_ary method, and only then assumed it was an array

In Screwcap, the Task class overrode method_missing, so when
Array#flatten called #to_ary, it performed the default method_missing
action for a Task object, which isn't quite what we want.

This commit makes :to_ary a special case and raises the appropriate
NoMethodError